TIMER BASICALLY WORKS VIA STATE PASSING
-bit of a back and forth with game and timer components, but it works
- maybe the use Ref would work better for different gameOver conditions
-Now I have to pass two functions for bot gameOver conditions to Timer:
Win and Timeout
Timer Counts when Time is out and activates the timeout function in game component,
wich sets the timer off, which renders the timer again to check if the timer is off,
and then the timer checks if the time has ran out or not and then chooses th correct
function, which is the time out(loss), to run in game component. Messy, I know.

the winning condition is more staright worwar because it happens first in the game component.
its just activates when all the characters are killed and stops the toggle Timer, which renders
the timer, which check again the correct condition, which in this case is the win condition because there
is still time left in the timer, and then runs the game components win-function